+++ include/cluaupp/libs/janitor.hpp
@@ -10,6 +10,8 @@ public:
 	template <typename T>
 	T Add(T object);
 	template <typename T>
+	T Add(T object, bool callDirectly);
+	template <typename T>
 	T Add(T object, string methodName);
 	template <typename T>
 	T Add(T object, string methodName, string index);
@@ -17,14 +19,15 @@ public:
 	T AddObject(T constructor);
 	template <typename P>
 	P AddPromise(P promise);
+	template <typename P>
+	P AddPromise(P promise, string index);
 	Janitor* Remove(string index);
 	Janitor* RemoveNoClean(string index);
 	Janitor* RemoveList(string index);
+	Janitor* RemoveListNoClean(string index);
 	template <typename T>
 	T Get(string index);
-	template <typename T>
-	T GetAll();
-	bool Has(string index);
+	void GetAll();
 	void Cleanup();
 	void Destroy();
 	RBXScriptConnection LinkToInstance(Instance* object);
@@ -32,4 +35,5 @@ public:
 	Janitor* LinkToInstances(Instance* object);
 	static bool Is(Instance* value);
 	static bool Is(string value);
+	static bool instanceof(Instance* value);
 };